requirements.txt
allure-pytest==2.5.2
pytest==3.9.1
pytest-rerunfailures==4.2
selenium==3.141.0
webdriver-manager==1.8.2
